+++ src/components/First.jsx
@@ -22,8 +22,15 @@
  * SOFTWARE.
  */
 import React, { Component } from 'react';
+import {
+  Button,
+  Col,
+  Form,
+  Row,
+} from 'react-bootstrap';
 
 import MatrixCanvas from './MatrixCanvas';
+import WSTable from './WSTable';
 
 /**
  * The component consists of the three main parts:
@@ -232,38 +239,51 @@ class First extends Component {
       messages,
       sessionId,
     } = this.state;
-    const messagesHtml = messages.map((message) => (
-      <li>
-        {message.author}
-        {': '}
-        {message.data}
-      </li>
-    ));
     return (
       <div>
-        <h3>Task A</h3>
-        <form>
-          <label htmlFor={this.sessionId}>
-          Session ID:
-            <input
-              ref={(component) => { this.sessionId = component; }}
+        <h1>
+          Task 1
+          <small className="text-muted">Digits recognition</small>
+        </h1>
+        <Row>
+          <Col xs={12} lg={6}>
+            <Form>
+              <Form.Row className="justify-content-md-center">
+                <Col xs={8}>
+                  <Form.Control
+                    ref={(component) => { this.sessionId = component; }}
+                    placeholder="Session ID"
+                  />
+                </Col>
+                <Col xs={4}>
+                  <Button
+                    variant="primary"
+                    type="button"
+                    onClick={() => this.observeSession()}
+                  >
+                    Observe
+                  </Button>
+                </Col>
+              </Form.Row>
+            </Form>
+          </Col>
+        </Row>
+        <Row className="justify-content-md-center">
+          <Col xs={6} lg={4}>
+            <MatrixCanvas
+              height={Math.round(First.HEIGHT)}
+              width={Math.round(First.WIDTH)}
+              matrix={matrix}
+              palette={{
+                0: '#FFFFFF',
+                1: '#000000',
+              }}
             />
-          </label>
-          <button type="button" onClick={() => this.observeSession()}>
-          Observe
-          </button>
-        </form>
-        <MatrixCanvas
-          height={Math.round(First.HEIGHT)}
-          width={Math.round(First.WIDTH)}
-          matrix={matrix}
-          palette={{
-            0: '#FFFFFF',
-            1: '#000000',
-          }}
-        />
-        <div>{sessionId ? `Session ${sessionId}` : ''}</div>
-        <ul>{messagesHtml}</ul>
+          </Col>
+        </Row>
+        <Row>
+          <WSTable messages={messages} sessionId={sessionId} />
+        </Row>
       </div>
     );
   }